Update how file/folderpickers work
