+++ StoryCADLib/Services/Dialogs/RecentFiles.xaml.cs
@@ -30,7 +30,7 @@ public sealed partial class RecentFiles : Page
                     StackPanel Item = new();
                     ToolTipService.SetToolTip(Item,File);
                     Item.Width = 300;
-                    Item.Children.Add(new TextBlock { Text = Path.GetFileName(File).Replace(".stbx", ""), FontSize = 20 });
+                    Item.Children.Add(new TextBlock { Text = Path.GetFileNameWithoutExtension(File), FontSize = 20 });
                     Item.Children.Add(new TextBlock { Text = "Last edited: " + System.IO.File.GetLastWriteTime(File), FontSize = 10, VerticalAlignment = VerticalAlignment.Center });
                     Recents.Items.Add(Item);
                 }
